requirements.txt
requests
python-dotenv
pandas
numpy
sqlalchemy
streamlit
plotly
